Refactor AgentLoop and ContextBuilder to enhance configuration management and concurrency handling. Introduced AgentDefaults for default settings, updated memory management with Arc and Mutex for thread safety, and added MessageBuilder trait for message construction. Implemented MessageTool for sending messages with optional media. Updated related configurations and tests to reflect these changes.

+++ src/agent/runner.rs
@@ -52,7 +52,7 @@ pub struct AgentRunSpec {
     pub fail_on_tool_error: bool,
     pub workspace: Option<PathBuf>,
     pub session_key: Option<String>,
-    pub context_window_tokens: Option<usize>,
+    pub context_window_tokens: Option<u32>,
     pub context_block_limit: Option<usize>,
     pub provider_retry_mode: String,
     pub progress_callback: Option<Arc<dyn Fn(Value) + Send + Sync>>,
@@ -315,7 +315,7 @@ impl AgentRunner {
 
         let max_output = spec.max_tokens.unwrap_or(4096);
         let budget = spec.context_block_limit.unwrap_or_else(|| {
-            context_window_tokens
+            (context_window_tokens as usize)
                 .saturating_sub(max_output)
                 .saturating_sub(SNIP_SAFETY_BUFFER)
         });
